+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # AutomationLabs Changelog
 
+## v0.1.2
+
+* Add Systems cli.
+* Add model cli :create for user models.
+* Add Github Actions with continuous integration yml.
+* Improve tests for continuous integration.
+
 ## v0.1.1
 
-* documentation with documenter.jl
+* Documentation with documenter.jl.
 
 ## v0.1.0